+++ CHANGELOG.md
@@ -1,3 +1,13 @@
 # Change Log
 
 ## [Unreleased]
+
+## [v0.2] - 2022-04-16
+### Changed
+- Return values from `encode()` and `decode()` rather a reference
+
+### Removed
+- `const char*` no longer supported as string
+
+[Unreleased]: https://github.com/terrakuh/smart_json/compare/v0.2..dev
+[v0.2]: https://github.com/terrakuh/smart_json/compare/v0.1..v0.2